pyproject.toml
[project]
name = "barcelona-factoids"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.52.0",
    "atproto>=0.0.61",
    "openai>=1.82.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "sqlglot[rs]>=26.22.0",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
    "ipython>=9.2.0",
]
